requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
streamlit>=1.43.0
